# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "Task app",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "This is a documentation of task app"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    },
    {
      "url": "task-pro-backend-mcfs.onrender.com"
    }
  ],
...
  "paths": {
...
    "/boards": {
      "get": {
        "tags": [
          "Boards"
        ],
        "summary": "Get all boards",
        "operationId": "getAllBoards",
        "description": "Retrieve all boards associated with the authenticated user.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved all boards",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully found boards"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "_id": {
                            "type": "string",
                            "example": "607d1c2e1f1b2c001c8e4d6b"
                          },
                          "title": {
                            "type": "string",
                            "example": "Project Planning"
                          },
                          "icon": {
                            "type": "object",
                            "properties": {
                              "name": {
                                "type": "string",
                                "example": "2_icon-star"
                              },
                              "iconUrl": {
                                "type": "string",
                                "format": "uri",
                                "example": "https://example.com/icons/2_icon-star.png"
                              }
                            }
                          },
                          "backgrounds": {
                            "type": "object",
                            "nullable": true,
                            "properties": {
                              "name": {
                                "type": "string",
                                "example": "bg-3"
                              },
                              "modalUrl": {
                                "type": "string",
                                "format": "uri",
                                "example": "https://modal.png"
                              },
                              "resolution": {
                                "type": "object",
                                "properties": {
                                  "mobile": {
                                    "type": "string",
                                    "format": "uri",
                                    "example": "https://mobile.jpg"
                                  },
                                  "tablet": {
                                    "type": "string",
                                    "format": "uri",
                                    "example": "https://tablet.jpg"
                                  },
                                  "desktop": {
                                    "type": "string",
                                    "format": "uri",
                                    "example": "https://desktop.jpg"
                                  }
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          }
        }
      },
      "post": {
        "tags": [
          "Boards"
        ],
        "summary": "Create a new board",
        "operationId": "createBoard",
        "description": "Create a new board for the authenticated user.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "title",
                  "iconName"
                ],
                "properties": {
                  "title": {
                    "type": "string",
                    "description": "The title of the board",
                    "example": "New Project Board"
                  },
                  "iconName": {
                    "type": "string",
                    "description": "The icon name for the board. Must be one of the predefined values.",
                    "enum": [
                      "1_icon-project",
                      "2_icon-star",
                      "3_icon-loading",
                      "4_icon-puzzle-piece",
                      "5_icon-container",
                      "6_icon-lightning",
                      "7_icon-colors",
                      "8_icon-hexagon"
                    ],
                    "example": "2_icon-star"
                  },
                  "backgroundName": {
                    "type": "string",
                    "description": "The background name for the board. Must be one of the predefined values.",
                    "enum": [
                      "bg-1",
                      "bg-2",
                      "bg-3",
                      "bg-4",
                      "bg-5",
                      "bg-6",
                      "bg-7",
                      "bg-8",
                      "bg-9",
                      "bg-10",
                      "bg-11",
                      "bg-12",
                      "bg-13",
                      "bg-14",
                      "bg-15"
                    ],
                    "example": "bg-3"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully created a new board",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully created a board"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "_id": {
                          "type": "string",
                          "example": "board ID (24 characters)"
                        },
                        "title": {
                          "type": "string",
                          "example": "New Project Board"
                        },
                        "icon": {
                          "type": "object",
                          "properties": {
                            "name": {
                              "type": "string",
                              "example": "2_icon-star"
                            },
                            "iconUrl": {
                              "type": "string",
                              "format": "uri",
                              "example": "https://example.com/icons/2_icon-star.png"
                            }
                          }
                        },
                        "backgrounds": {
                          "type": "object",
                          "nullable": true,
                          "properties": {
                            "name": {
                              "type": "string",
                              "example": "bg-3"
                            },
                            "modalUrl": {
                              "type": "string",
                              "format": "uri",
                              "example": "https://modal.png"
                            },
                            "resolution": {
                              "type": "object",
                              "properties": {
                                "mobile": {
                                  "type": "string",
                                  "format": "uri",
                                  "example": "https://mobile.jpg"
                                },
                                "tablet": {
                                  "type": "string",
                                  "format": "uri",
                                  "example": "https://tablet.jpg"
                                },
                                "desktop": {
                                  "type": "string",
                                  "format": "uri",
                                  "example": "https://desktop.jpg"
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Missing or invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 400
                    },
                    "message": {
                      "type": "string",
                      "example": "Invalid request parameters"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          }
        }
      }
    },
    "/boards/info": {
      "get": {
        "tags": [
          "Boards"
        ],
        "summary": "Retrieve Icons and Backgrounds",
        "operationId": "getBackgroundsAndIcons",
        "description": "Retrieves a list of available icons and backgrounds for the boards. \nReturns the details including their IDs, names, and URLs.\n",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved icons and backgrounds",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully found icons and backgrounds"
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "icons",
                        "backgrounds"
                      ],
                      "properties": {
                        "icons": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "required": [
                              "_id",
                              "name",
                              "iconUrl"
                            ],
                            "properties": {
                              "_id": {
                                "type": "string",
                                "description": "Unique identifier for the icon",
                                "example": "672543634967d1183bbad202"
                              },
                              "name": {
                                "type": "string",
                                "description": "Name of the icon",
                                "example": "1_icon-project"
                              },
                              "iconUrl": {
                                "type": "string",
                                "format": "uri",
                                "description": "URL of the icon image",
                                "example": "https://res.cloudinary.com/dxlahrm0m/image/upload/v1730495328/icons/exet0qmduhppm344sh5x.svg"
                              }
                            }
                          }
                        },
                        "backgrounds": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "required": [
                              "_id",
                              "name",
                              "modalUrl"
                            ],
                            "properties": {
                              "_id": {
                                "type": "string",
                                "description": "Unique identifier for the background",
                                "example": "6724dc14e9a3f1f8b9746b3b"
                              },
                              "name": {
                                "type": "string",
                                "description": "Name of the background",
                                "example": "bg-1"
                              },
                              "modalUrl": {
                                "type": "string",
                                "format": "uri",
                                "description": "URL of the background image for modal",
                                "example": "https://res.cloudinary.com/dxlahrm0m/image/upload/v1730468881/backgrounds/nbxdi3fvgb6rjhp9uxca.png"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Icons or backgrounds not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 404
                    },
                    "message": {
                      "type": "string",
                      "example": "Icons not found"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
    "responses": {
      "401": {
        "description": "Unauthorized error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 401
                },
                "message": {
                  "type": "string",
                  "example": "UnauthorizedError"
                },
                "data": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Access token expired"
                    }
                  }
                }
              }
            }
          }
        }
      },
...
    }
  }
}